Update damage plate textures for consistency
The damage plate texture has been updated to include the square lines that
are center to the design of the other damage plates. The color has also
been changed to be closer to the crimson fungus which is a primary
crafting ingredient. The damage plate has been slightly changed to the
same shade of cyan as warped fungus for the same reason.

The maim plate has been changed to be more purple in color. This is to
further differentiate it from the flame plate which had effectively the
same color as the previous design.